# Doc/swagger.yaml
openapi: 3.0.3
info:
  title: Green Dii API
  description: |
    REST API for Green Dii Sales Workflow System
    ระบบจัดการ Sales Workflow — Quotation → Work Order → Hand Over → Purchase Request
    **Default password for all seeded users: `1234`**
  version: 1.0.0
  contact:
    name: Green Dii Support
servers:
  - url: http://localhost:4000/api
    description: Development
  - url: https://your-domain.com/api
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WorkOrder:
      type: object
      properties:
        id:           { type: string }
        woNo:         { type: string }
        quotationId:  { type: string }
        project:      { type: string }
        customerName: { type: string }
        status:       { type: string, enum: [draft, pending, approved, rejected] }
        approvalStep: { type: integer }
        isClosed:     { type: boolean }
        createdAt:    { type: string, format: date-time }
...
paths:
...
  /workorders:
    get:
      tags: [WorkOrders]
      summary: List work orders
      responses:
        "200": { description: Work order list }
    post:
      tags: [WorkOrders]
      summary: Create work order
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrder' }
      responses:
        "201": { description: Created }
  /workorders/{id}/approve:
    post:
      tags: [WorkOrders]
      summary: Approve work order step
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment: { type: string }
      responses:
        "200": { description: Approved }
  /workorders/{id}/reject:
    post:
      tags: [WorkOrders]
      summary: Reject work order
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Rejected }

  # ── HAND OVER JOBS ────────────────────────────────────────────────────────────
...